healthcare-data/requirements.txt
# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9

# Environment and API
python-dotenv==1.0.0
requests==2.31.0

# Data processing
numpy==1.24.3
torch==2.0.1
sentence-transformers==2.2.2
huggingface-hub==0.16.4
